refactor: implement modular on api token
